move variant into main list file
